interstats for openbsd.

+++ bsd/intrstats.h
@@ -47,7 +47,7 @@ template <class X>
 std::map<size_t, uint64_t> IntrStats::getCounts(const std::vector<X> &v) const {
     std::map<size_t, uint64_t> rval;
 
-    std::map<size_t, uint64_t>::const_iterator it;
+    std::map<size_t, size_t>::const_iterator it;
     for (it = _irqMap.begin() ; it != _irqMap.end() ; ++it)
         rval[it->first] = v[it->second];
